Fix px-app-route bugs impacting Polymer 2 apps
Polymer 2.x changes to lifecycle timing and observer firing (mainly
that observers now fire before all arguments are undefined) caused
px-app-route to fail in Polymer 2.x in multiple ways. It would not
decode the initial route, would not update the route, and would
break routing in apps by resetting a valid route to an empty
route when the user first loaded the page.

This change fixes that and makes px-app-route work the same in
Polymer 1.x and 2.x. It also adds some acceptance tests to try
and ensure a bug like this does not happen in the future.