Fix property set order bug in px-app-asset-selectable behavior
In Polymer 2, properties are decoded to default values at a different
time than in Polymer 1.x. In Polymer 2 the the asset graph can be
created early, before the selectedRoute property has been evaluated
to its default null value. This creates a condition where we try to
call into our asset graph to get an asset that does not exist. This
fix ensures we do not throw an error and stop rendering by stopping
execution if we hit this case.