Port changed code from selected behavior back to activatable behavior
The intention was always to keep the px-app-asset-selectable-behavior
and the px-app-asset-activatable behavior in sync with the same
code but different names for everything. That is hard to do and
some refactoring in selectable didn't make it to activatable.
That lead to a race condition where the asset graph was ready
before we expected to appear in activatable. It turns out we had
already solved that problem in selectable. I ported the fix over
so things should work as expected.